Attempt workaround for issue with Windows Server 2016 baseline image
The latest attempt to build the Windows Server 2016 baseline image hangs (repeatedly) while installing the December 2024 Cumulative Update (CU).

To avoid the issue, add packages to slipstream the latest Servicing Stack Update (SSU) and CU for Windows Server 2016.

Note: It turns out this doesn't work either, because DISM can't apply the SSU and the CU in a single step. Instead, it throws an error stating the update (presumably the CU) doesn't apply.

References:

- https://www.reddit.com/r/MDT/comments/axji7s/injecting_ssus_before_dependent_cus/
- https://damgoodadmin.com/2018/12/20/servicing-stack-updates-what-is-this-madness/